♻️ refactor ScrollTopButton
