+++ src/components/uiParts/index.ts
@@ -5,5 +5,6 @@ export { Input } from './Input';
 export { Link } from './Link';
 export { Loading } from './Loading';
 export { LoginCard } from './LoginCard';
+export { ScrollTopButton } from './ScrollTopButton';
 export { Text } from './Text';
 export { Tooltip } from './Tooltip';
